food_delivery_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
geoalchemy2
psycopg2-binary
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
